update demo page for Grid Component
