+++ misc/demo/app/DemoApp.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Router, Route, browserHistory } from 'react-router';
 
 import Page from './Page';
-import GridSystem from './GridSystem';
+import GridSystem from './pages/GridSystem';
 import PageHeader from './pages/PageHeader';
 import Icons from './pages/Icon';
 import MediaObject from './pages/MediaObject';
@@ -14,8 +14,6 @@ const DemoApp = () =>
     <Route path="/" component={Page}>
       <Route path="/grid-system" component={GridSystem} />
       <Route path="/icons" component={Icons} />
-      <Route path="/media-objects" component={MediaObjects} />
-      <Route path="/page-headers" component={PageHeaders} />
       <Route path="/buttons" component={Buttons} />
       <Route path="/badges" component={Badges} />
       <Route path="/page-headers" component={PageHeader} />
