+++ misc/demo/app/pages/GridSystem/ExampleManual.js
@@ -1,0 +1,34 @@
+import React from 'react';
+
+import { Grid, Column, Box } from 'react-lds';
+
+const ExampleManual = () =>
+  <Grid pull-padded wrap>
+    <Column className="slds-m-bottom--large" padded size-of="1-1">
+      <Box theme="shade">1</Box>
+    </Column>
+    <Column className="slds-m-bottom--large" padded size-of="1-2" medium-size-of="5-6" large-size-of="8-12">
+      <Box theme="shade">2</Box>
+    </Column>
+    <Column className="slds-m-bottom--large" padded size-of="1-2" medium-size-of="1-6" large-size-of="4-12">
+      <Box theme="shade">3</Box>
+    </Column>
+    <Column className="slds-m-bottom--large" padded size-of="1-1" medium-size-of="1-2" large-size-of="1-3">
+      <Box theme="shade">4</Box>
+    </Column>
+    <Column className="slds-m-bottom--large" padded size-of="1-1" medium-size-of="1-2" large-size-of="1-3">
+      <Box theme="shade">5</Box>
+    </Column>
+    <Column className="slds-m-bottom--large" padded size-of="1-1" large-size-of="1-3">
+      <Grid pull-padded wrap>
+        <Column className="slds-m-bottom--large" padded size-of="1-2" medium-size-of="1-1" large-size-of="1-2">
+          <Box theme="shade">6</Box>
+        </Column>
+        <Column className="slds-m-bottom--large" padded size-of="1-2" medium-size-of="1-1" large-size-of="1-2">
+          <Box theme="shade">7</Box>
+        </Column>
+      </Grid>
+    </Column>
+  </Grid>;
+
+export default ExampleManual;
